Split my TLK hierarchy to allow for future code reuse.
