+++ DB/src/main/java/io/deephaven/db/v2/locations/local/FileTableLocationKey.java
@@ -1,0 +1,89 @@
+package io.deephaven.db.v2.locations.local;
+
+import io.deephaven.base.log.LogOutput;
+import io.deephaven.db.v2.locations.ImmutableTableLocationKey;
+import io.deephaven.db.v2.locations.PartitionedTableLocationKey;
+import io.deephaven.db.v2.locations.TableLocationKey;
+import io.deephaven.io.log.impl.LogOutputStringImpl;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.io.File;
+import java.util.Map;
+
+/**
+ * Base {@link ImmutableTableLocationKey} implementation for table locations that may be enclosed by partitions and
+ * described by a {@link File}.
+ * Sub-classes should override {@link #compareTo(TableLocationKey)} and {@link #equals(Object)} only if they need to
+ * prevent equality with other {@link FileTableLocationKey} implementations.
+ */
+public abstract class FileTableLocationKey extends PartitionedTableLocationKey {
+
+    protected final File file;
+
+    private int cachedHashCode;
+
+    /**
+     * Construct a new FileTableLocationKey for the supplied {@code file} and {@code partitions}.
+     *
+     * @param file        The file (or directory) that backs the keyed location. Will be adjusted to an absolute path.
+     * @param partitions The table partitions enclosing the table location keyed by {@code this}. Note that if this
+     *                   parameter is {@code null}, the location will be a member of no partitions. An ordered copy
+     *                   of the map will be made, so the calling code is free to mutate the map after this call
+     *                   completes, but the partition keys and values themselves <em>must</em> be effectively immutable.
+     */
+    public FileTableLocationKey(@NotNull final File file, @Nullable final Map<String, Comparable<?>> partitions) {
+        super(partitions);
+        this.file = file.getAbsoluteFile();
+    }
+
+    @Override
+    public LogOutput append(@NotNull final LogOutput logOutput) {
+        return logOutput.append(getImplementationName())
+                .append(":[file=").append(file.getPath())
+                .append(",partitions=").append(PartitionsFormatter.INSTANCE, partitions)
+                .append(']');
+    }
+
+    @Override
+    public String toString() {
+        return new LogOutputStringImpl().append(this).toString();
+    }
+
+    @Override
+    public int compareTo(@NotNull final TableLocationKey other) {
+        if (other instanceof FileTableLocationKey) {
+            final FileTableLocationKey otherTyped = (FileTableLocationKey) other;
+            final int partitionComparisonResult = PartitionsComparator.INSTANCE.compare(partitions, otherTyped.partitions);
+            return partitionComparisonResult != 0 ? partitionComparisonResult : file.compareTo(otherTyped.file);
+        }
+        throw new ClassCastException("Cannot compare " + getClass() + " to " + other.getClass());
+    }
+
+    @Override
+    public int hashCode() {
+        if (cachedHashCode == 0) {
+            final int computedHashCode = 31 * partitions.hashCode() + file.hashCode();
+            // Don't use 0; that's used by StandaloneTableLocationKey, and also our sentinel for the need to compute
+            if (computedHashCode == 0) {
+                final int fallbackHashCode = FileTableLocationKey.class.hashCode();
+                cachedHashCode = fallbackHashCode == 0 ? 1 : fallbackHashCode;
+            } else {
+                cachedHashCode = computedHashCode;
+            }
+        }
+        return cachedHashCode;
+    }
+
+    @Override
+    public boolean equals(@Nullable final Object other) {
+        if (this == other) {
+            return true;
+        }
+        if (!(other instanceof FileTableLocationKey)) {
+            return false;
+        }
+        final FileTableLocationKey otherTyped = (FileTableLocationKey) other;
+        return file.equals(otherTyped.file) && partitions.equals(otherTyped.partitions);
+    }
+}
